+++ CHANGES.rst
@@ -5,3 +5,7 @@
 
 - #2612401: Don't fail if the user can't be found.
   [JL 2019-02-28]
+
+- #2745054: Show Notifications menu item if there are read notifications and no
+  unread notifications.
+  [JL 2019-05-29]